+++ Documentation/ChangeLog.md
@@ -4,6 +4,9 @@
 
 ### Added
 
+- Made it so all upgrade buttons dynamically show up according to their price **(Andrew)**
+- Fixed event lock for gambling event **(Andrew)**
+- Fixed display bug of upgrades prices **(Andrew)**
 - Support for multiple save slots **(Jack)**
 
 ### Modified